UF-14339: Fix sonar issue
